Implement a polling + internal queue mechanism to update the gui (isseu #8).
The previous threading model caused the gui to hang.  That was likely due to the
fact that not all the gui updates where done by the mainloop thread.  Now, we
have a periodic job which looks for pending gui operations and schedule
functions if necessary.  Issue #8 should be fixed now.